+++ oblate/fields/typings.py
@@ -1,0 +1,46 @@
+# MIT License
+
+# Copyright (c) 2023 Izhar Ahmad
+
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+from __future__ import annotations
+
+from typing import Any, TYPE_CHECKING
+from oblate.fields.base import Field
+
+if TYPE_CHECKING:
+    from oblate.contexts import LoadContext, DumpContext
+
+__all__ = (
+    'Any',
+)
+
+
+class Any(Field[Any, Any]):
+    """A field that accepts any arbitrary value.
+
+    This field acts as a "raw field" that performs no validation on the
+    given value.
+    """
+    def value_load(self, value: Any, context: LoadContext) -> Any:
+        return value
+
+    def value_dump(self, value: Any, context: DumpContext) -> Any:
+        return value